+++ src/Chunkyard.Tests/Core/SerializeTests.cs
@@ -21,15 +21,15 @@ public static class SerializeTests
     }
 
     [Fact]
-    public static void Serialize_Can_Convert_LogReference()
+    public static void Serialize_Can_Convert_SnapshotReference()
     {
-        var expected = new LogReference(
+        var expected = new SnapshotReference(
             Crypto.GenerateSalt(),
             Crypto.DefaultIterations,
             new[] { "ad95131bc0b799c0b1af477fb14fcf26a6a9f76079e48bf090acb7e8367bfd0e" });
 
-        var actual = Serialize.BytesToLogReference(
-            Serialize.LogReferenceToBytes(expected));
+        var actual = Serialize.BytesToSnapshotReference(
+            Serialize.SnapshotReferenceToBytes(expected));
 
         Assert.Equal(expected, actual);
     }